+++ app/helpers/sheets_helper.rb
@@ -12,8 +12,19 @@ module SheetsHelper
 
   def filter_link(count, design, variable, value)
     params = { design_id: design.id }
-    params[:f] = [{ variable_id: variable.id, value: value }] if variable
+    params[:search] = "#{variable.name}:#{convert_value(value)}" if variable
     url = project_sheets_path design.project, params
     link_to_if count.present?, count || '-', url, target: '_blank'
+  end
+
+  def convert_value(value)
+    case value
+    when ':any'
+      'any'
+    when ':missing'
+      'missing'
+    else
+      value
+    end
   end
 end